Better separate generated and handcrafted files
