+++ lib/src/com/IUnknown.dart
@@ -3,9 +3,8 @@ import 'dart:ffi';
 
 import 'package:ffi/ffi.dart';
 
-import 'combase.dart';
-import 'comerrors.dart';
-
+import '../combase.dart';
+import '../comerrors.dart';
 import '../constants.dart';
 import '../macros.dart';
 import '../structs.dart';
