+++ lib/src/com/IWbemServices.dart
@@ -4,9 +4,8 @@ import 'dart:ffi';
 import 'package:ffi/ffi.dart';
 
 import 'IUnknown.dart';
-import 'combase.dart';
-import 'comerrors.dart';
-
+import '../combase.dart';
+import '../comerrors.dart';
 import '../constants.dart';
 import '../macros.dart';
 import '../structs.dart';
